refactor kindnetd
